backend/pyproject.toml
[tool.poetry]
name = "ai-positivity-transformer"
version = "0.1.0"
description = ""
authors = ["Asako Hayase <asako.ch.ha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-multipart = "^0.0.17"
requests = "^2.32.3"
pydantic = "^2.10.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
isort = "^5.13.2"
pre-commit = "^4.0.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn src.main:app --reload"
format = "black ."
sort = "isort ."
lint = "flake8"
typecheck = "mypy ."
test = "pytest"
